calculate keyexpantion with mpc seperated prover_lib and verifier_lib from common
